split code into library, admin and api
