+++ .idea/modules.xml
@@ -2,7 +2,9 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
+      <module fileurl="file://$PROJECT_DIR$/.idea/modules/admin/quran4j.admin.main.iml" filepath="$PROJECT_DIR$/.idea/modules/admin/quran4j.admin.main.iml" />
       <module fileurl="file://$PROJECT_DIR$/.idea/modules/api/quran4j.api.main.iml" filepath="$PROJECT_DIR$/.idea/modules/api/quran4j.api.main.iml" />
+      <module fileurl="file://$PROJECT_DIR$/.idea/modules/library/quran4j.library.main.iml" filepath="$PROJECT_DIR$/.idea/modules/library/quran4j.library.main.iml" />
     </modules>
   </component>
 </project>
