+++ admin/src/main/java/quran4j/admin/AdminApplication.java
@@ -1,17 +1,17 @@
-package quran4j.api;
+package quran4j.admin;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
-@SpringBootApplication(scanBasePackages = { "quran4j.library", "quran4j.api" })
+@SpringBootApplication(scanBasePackages = { "quran4j.library", "quran4j.admin" })
 @EnableJpaRepositories(basePackages = "quran4j.library")
 @EntityScan(basePackages = "quran4j.library")
-public class ApiApplication {
+public class AdminApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(ApiApplication.class, args);
+		SpringApplication.run(AdminApplication.class, args);
 	}
 
 }
