+++ admin/src/main/java/quran4j/admin/dto/AyatDTO.java
@@ -1,0 +1,4 @@
+package quran4j.admin.dto;
+
+public record AyatDTO(int chapter, int verse, String text) {
+}
