+++ admin/src/main/java/quran4j/admin/dto/SurahDTO.java
@@ -1,0 +1,4 @@
+package quran4j.admin.dto;
+
+public record SurahDTO(int id, String translation, String type) {
+}
