+++ api/src/main/java/quran4j/api/ayat/dto/AyatDTO.java
@@ -1,0 +1,4 @@
+package quran4j.api.ayat.dto;
+
+public record AyatDTO(int index, String text) {
+}
