--- api/src/main/java/quran4j/api/dto/AyatDTO.java
@@ -1,4 +1,0 @@
-package quran4j.api.dto;
-
-public record AyatDTO(int chapter, int verse, String text) {
-}
