--- api/src/main/java/quran4j/api/dto/QuranDTO.java
@@ -1,7 +1,0 @@
-package quran4j.api.dto;
-
-import java.util.List;
-import java.util.Map;
-
-public record QuranDTO(Map<String, List<AyatDTO>> ayatsBySurahIndex) {
-}
