--- api/src/main/java/quran4j/api/dto/SurahChapterDTO.java
@@ -1,4 +1,0 @@
-package quran4j.api.dto;
-
-public record SurahChapterDTO(int id, String translation, String type) {
-}
