--- api/src/main/java/quran4j/api/model/Ayat.java
@@ -1,11 +1,0 @@
-package quran4j.api.model;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public class Ayat {
-    private final int index;
-    private final String text;
-}
