--- api/src/main/java/quran4j/api/service/QuranService.java
@@ -1,7 +1,0 @@
-package quran4j.api.service;
-
-import quran4j.api.model.Quran;
-
-public interface QuranService {
-    Quran save(Quran quran);
-}
