+++ library/src/main/java/quran4j/library/ayat/persistence/mapper/AyatDomainMapper.java
@@ -1,0 +1,10 @@
+package quran4j.library.ayat.persistence.mapper;
+
+import org.mapstruct.Mapper;
+import quran4j.library.ayat.persistence.domain.AyatDomain;
+import quran4j.library.ayat.business.model.Ayat;
+
+@Mapper(componentModel = "spring")
+public interface AyatDomainMapper {
+    AyatDomain toDomain(Ayat model);
+}
