+++ settings.gradle
@@ -6,4 +6,4 @@
  */
 
 rootProject.name = 'quran4j'
-include("api")
+include("api", "admin", "library")
